update translations and  move util function into hook folder
