+++ lib/Capabilities/CheckboxWithAsterisk/CheckboxWithAsterisk.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import { Checkbox } from '@folio/stripes/components';
 
-import css from './index.css';
+import css from './style.css';
 
 export const CheckboxWithAsterisk = ({ count, ...rest }) => {
   return (
