+++ lib/Capabilities/CheckboxWithAsterisk/style.css
