+++ lib/Capabilities/ItemActionCheckbox/ItemActionCheckbox.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { NoValue } from '@folio/stripes/components';
 
 import { CheckboxWithAsterisk } from '../CheckboxWithAsterisk';
-import { useCheckboxAriaStates } from '../utils';
+import { useCheckboxAriaStates } from './hooks';
 
 export const ItemActionCheckbox = ({
   item,
