+++ lib/Capabilities/ItemActionCheckbox/hooks/index.js
@@ -1,0 +1,1 @@
+export { useCheckboxAriaStates } from './useCheckboxAriaStates';
