+++ lib/Capabilities/ItemActionCheckbox/hooks/useCheckboxAriaStates.js
@@ -1,5 +1,6 @@
 import { useIntl } from 'react-intl';
-import { columnTranslations } from '../../constants/translations';
+
+import { columnTranslations } from '../../constants';
 
 export const useCheckboxAriaStates = () => {
   const { formatMessage } = useIntl();
@@ -8,4 +9,3 @@ export const useCheckboxAriaStates = () => {
 
   return { getCheckboxAriaLabel };
 };
-
