+++ lib/Capabilities/constants.js
@@ -46,96 +46,78 @@ export const visibleColumns = [
 
 export const columnTranslations = defineMessage({
   'name': {
-    id: 'ui-authorization-roles.columns.name',
+    id: 'stripes-authorization-components.columns.name',
     defaultMessage: 'Name',
     description: 'Name column label',
   },
   'description': {
-    id: 'ui-authorization-roles.columns.description',
+    id: 'stripes-authorization-components.columns.description',
     defaultMessage: 'Description',
     description: 'Description column label',
   },
   'updatedDate': {
-    id: 'ui-authorization-roles.columns.updatedDate',
+    id: 'stripes-authorization-components.columns.updatedDate',
     defaultMessage: 'Updated',
     description: 'Updated date column label',
   },
   'updatedBy': {
-    id: 'ui-authorization-roles.columns.updatedBy',
+    id: 'stripes-authorization-components.columns.updatedBy',
     defaultMessage: 'Updated by',
     description: 'Updated by column label',
   },
   [CAPABILITIES_COLUMN_NAMES.application]: {
-    id: 'ui-authorization-roles.columns.application',
+    id: 'stripes-authorization-components.columns.application',
     defaultMessage: 'Application',
     description: 'Application column label',
   },
   [CAPABILITIES_COLUMN_NAMES.resource]: {
-    id: 'ui-authorization-roles.columns.resource',
+    id: 'stripes-authorization-components.columns.resource',
     defaultMessage: 'Resource/Resource sets',
     description: 'Resource/Resource sets column label',
   },
   [CAPABILITIES_COLUMN_NAMES.view]: {
-    id: 'ui-authorization-roles.columns.view',
+    id: 'stripes-authorization-components.columns.view',
     defaultMessage: 'View',
     description: 'View column label',
   },
   [CAPABILITIES_COLUMN_NAMES.edit]: {
-    id: 'ui-authorization-roles.columns.edit',
+    id: 'stripes-authorization-components.columns.edit',
     defaultMessage: 'Edit',
     description: 'Edit column label',
   },
   [CAPABILITIES_COLUMN_NAMES.create]: {
-    id: 'ui-authorization-roles.columns.create',
+    id: 'stripes-authorization-components.columns.create',
     defaultMessage: 'Create',
     description: 'Create column label',
   },
   [CAPABILITIES_COLUMN_NAMES.delete]: {
-    id: 'ui-authorization-roles.columns.delete',
+    id: 'stripes-authorization-components.columns.delete',
     defaultMessage: 'Delete',
     description: 'Delete column label',
   },
   [CAPABILITIES_COLUMN_NAMES.manage]: {
-    id: 'ui-authorization-roles.columns.manage',
+    id: 'stripes-authorization-components.columns.manage',
     defaultMessage: 'Manage',
     description: 'Manage column label',
   },
   'policies': {
-    id: 'ui-authorization-roles.columns.policies',
+    id: 'stripes-authorization-components.columns.policies',
     defaultMessage: 'Policies',
     description: 'Policies column label',
   },
   'settings': {
-    id: 'ui-authorization-roles.columns.settings',
+    id: 'stripes-authorization-components.columns.settings',
     defaultMessage: 'Settings/Settings sets',
     description: 'Settings/Settings sets column label',
   },
   'procedure': {
-    id: 'ui-authorization-roles.columns.procedure',
+    id: 'stripes-authorization-components.columns.procedure',
     defaultMessage: 'Procedure',
     description: 'Procedure column label',
   },
   [CAPABILITIES_COLUMN_NAMES.execute]: {
-    id: 'ui-authorization-roles.columns.execute',
+    id: 'stripes-authorization-components.columns.execute',
     defaultMessage: 'Execute',
     description: 'Execute column label',
-  },
-});
-
-export const formTranslations = defineMessage({
-  name: {
-    id: 'ui-authorization-roles.form.labels.name',
-    defaultMessage: 'Name',
-    description: 'Name label in a form',
-  },
-  description: {
-    id: 'ui-authorization-roles.form.labels.description',
-    defaultMessage: 'Description',
-    description: 'Description label in a form',
-  },
-  multiplePolicies: {
-    id: 'ui-authorization-roles.form.labels.description',
-    defaultMessage: 'Description',
-    description: 'Description label in a form',
   },
 });
