--- lib/Capabilities/styles.css
@@ -1,31 +1,0 @@
-@import "@folio/stripes-components/lib/variables";
-
-.lookupSearchContainer {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.lookupSearch {
-  flex-grow: 1;
-  margin-right: var(--gutter-static-two-thirds);
-  width: unset;
-}
-
-.alignRightWrapper{
-  display: flex;
-  justify-content: flex-end;
-}
-
-.gridCheckbox {
-  margin-left: 24px;
-}
-
-.gutterTop{
-  margin-block-start: var(--gutter);
-}
-
-.loadingInTitle{
-  margin-left: 8px;
-}
