+++ lib/Capabilities/utils.js
@@ -1,17 +1,7 @@
-import { useIntl } from 'react-intl';
-
 import {
   columnTranslations,
   CAPABILITIES_COLUMN_NAMES,
 } from './constants';
-
-export const useCheckboxAriaStates = () => {
-  const { formatMessage } = useIntl();
-
-  const getCheckboxAriaLabel = (action, resource) => `${formatMessage(columnTranslations[action])} ${resource}`;
-
-  return { getCheckboxAriaLabel };
-};
 
 export const getResultsFormatter = (renderItemActionCheckbox) => ({
   [CAPABILITIES_COLUMN_NAMES.application]: (item) => item.applicationId,
