--- test/__mock__/index.js
@@ -1,7 +1,0 @@
-import './matchMedia.mock';
-import './stripesConfig.mock';
-import './stripesCore.mock';
-import './stripesIcon.mock';
-import './stripesComponents.mock';
-import './resizeObserver.mock';
-import './intl.mock';
