--- test/__mock__/intl.mock.js
@@ -1,34 +1,0 @@
-import React from 'react';
-
-jest.mock('react-intl', () => {
-  const intl = {
-    formatMessage: ({ id }) => id,
-  };
-
-  return {
-    ...jest.requireActual('react-intl'),
-    FormattedMessage: jest.fn(({ id, children }) => {
-      if (children) {
-        return children([id]);
-      }
-
-      return id;
-    }),
-    FormattedTime: jest.fn(({ value, children }) => {
-      if (children) {
-        return children([value]);
-      }
-
-      return value;
-    }),
-    FormattedDate: jest.fn(({ value, children }) => {
-      if (children) {
-        return children([value]);
-      }
-
-      return value;
-    }),
-    useIntl: () => intl,
-    injectIntl: (Component) => (props) => <Component {...props} intl={intl} />,
-  };
-});
