--- test/__mock__/resizeObserver.mock.js
@@ -1,5 +1,0 @@
-global.ResizeObserver = jest.fn().mockImplementation(() => ({
-  observe: jest.fn(),
-  unobserve: jest.fn(),
-  disconnect: jest.fn(),
-}));
