--- test/__mock__/stripesConfig.mock.js
@@ -1,1 +1,0 @@
-jest.mock('stripes-config', () => ({ modules: [], metadata: {} }), { virtual: true });
