--- test/__mock__/stripesIcon.mock.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-jest.mock('@folio/stripes-components/lib/Icon', () => {
-  return props => {
-    return props.children ? props.children : <span />;
-  };
-});
